Fix: gradcheck forced float32 (#8230)

+++ torch/autograd/gradcheck.py
@@ -21,7 +21,7 @@ def make_jacobian(input, num_out):
             return None
         if not input.requires_grad:
             return None
-        return torch.zeros(input.nelement(), num_out)
+        return torch.zeros(input.nelement(), num_out, dtype=input.dtype)
     elif isinstance(input, Iterable):
         jacobians = list(filter(
             lambda x: x is not None, (make_jacobian(elem, num_out) for elem in input)))
